Rename processors to schemes
